transformer/requirements.txt
paho-mqtt
datetime
python-dateutil
redis
confluent_kafka
kafka
